klte-common: add GPS overlay
xtra server url need to be set, it is empty
in framework, without those xtra server set,
it will not download the gps almanacs,
ephemerides data from wifi or radio network. It will
need to be download from gps satellites directly
which is very slow (> 10 minutes if download from a
single satellite).

Change-Id: Ide158e0095a6de1068d404b98fd4dc59608f4720

Conflicts:
	overlay/frameworks/base/core/res/res/values/config.xml